+++ src/lib.rs
@@ -89,6 +89,7 @@ pub mod stock_str;
 mod sync;
 mod timesmearing;
 mod token;
+mod transport;
 mod update_helper;
 pub mod webxdc;
 #[macro_use]
